Don't throw if build/ doesn't exist

+++ deploy/deploy.sh
@@ -2,7 +2,7 @@
 
 set -xeo pipefail
 
-rm -r build/
+[ -d build/ ] && rm -r build/
 yarn
 node lib/update-projects.js
 NODE_ENV=production yarn build
